==> 01_data_preprocessing.ipynb <==
# Créditos
Los bancos son fundamentales para el desarrollo del mercado. El crédito es una de las herramientas con mayor uso y su uso data desde el renacimiento. El concepto es sencillo, se obtiene un nominal de manera inmediata a cambio de pagar un interes por el tiempo y la cantidad que se haya obtenido. Sin embargo, por muy moderno que sean los sistemas actualmente, el credito sigue sustentado en un acto de fe. De hecho el nombre  proviene del latín del verbo credere, que significa confianza.

En tiempo mas recientes la manera de obtener un crédito era muy complicado y muy subjetivo. El cliente tenia que acudir a una institución bancaria en la cual un ejecutivo estudiaba si era candidato o no a un crédito, en este contexto tenemos una parte importante del crétido. Él término de reputación, este estudio era mas subjetivo que cuantitativo, se tomaba incluso la vestimenta y opiniones de personas cercanas. Con el avance de los sistemas estos métodos fueron sofisticados de manera paulatina.

Por ejemplo en México se cuenta con Buró de crédito, el cual contien el historial financiero del cliente, incluisive de activides cotidianas. Desde servicios telefonicos hasta compras departamentales.

# Problematica
Una institución bancaria, debe tener sólidos sistemas para el prestamo de capital. Técncias coomo XVA y CVA, son centrales para determinar que tanto puede perder la institución si un cliente no paga adecuadamente sus prestamos. Más aún, la cualidad de los créditos son muy importantes, por ejemplo meedir cuanto nos costaria que un cliente pague anticipadamene un crédito, pues recordemos que un interes está asociado al tiempo y a la cantidad prestada. Por otro lado, el central en este ejercicio, cuanto podemos perder como institución al ceder de manera erronea un crédito a un perfil riesgoso.

En términos comunes, nos interesa poder determinar la probabilidad de inpago de una contraparte, pues no solo es importe conocer el valor presente de un subyacente financiero, si no tambien su evolución.

Esto no es reciente pero si económicante hablando, las tecnicas son originadas desde la crisis suprime de 2008, en la cual se asumió erroneamente que las contrapartes siempre iban a pagar sus hipótecas, pero los sistemas no pudieron medir el valor de los CDO's asociados a esas hipotecas, lo cual culminó en la crisis inmobiliraria ya mencionada.

# Contexto
El data set, se obtubo de Kaggle,

https://www.kaggle.com/c/GiveMeSomeCredit

este data set, contiene informacion sobre clientes, la cual destaca su naturaliza que es interes al autor de este pre-estudio. Si bien es de interes conocer la morosidad  suelen ser mucho menores que los de "no default".

Se busca entrenar un modelo de XGBoostm, en este caso, nuestra intensión es lograr dominar ténicas de control del desbalanceo, el principal riesgo de ésta osadia es que el modelo podría sesgarse hacia la clase mayoritaria (prediciendo casi siempre "no default"), lo que no sería útil en la práctica.

Buscamos experiencia para el desbalanceo en el dataset con cosas como;
- Ajustar el parámetro scale_pos_weight;
- Sobremuestreo o submuestreo,
- Métricas alternativas;
  - ROC-AUC;
  - F1-Score;
  - Recall.
- Ajuste del umbral de decisión.

Actualmente se puede usar técnicas estadísticas como
[Bootstrapping ](https://en.wikipedia.org/wiki/Bootstrapping_(statistics)) para determinar las proyecciones del valor de los productos financieros. Sin embargo estos procesos son caros y en la práctica presenta errores a cambios de tasas, como recien se vio en los cambios de tasas de la FED de septiembre.

Buscamos un modelos que nos permita ahorrar en los sistemas de externos, que nos permita determinar el riesgo asociado a una contraparte. En este ejercicio se tiene la información financiera de indvioduos cuya parte dificil es determinar que un cliente con las condiciones presentadas sea o no un usuario moroso.

Lo importantes de este data set, es la cantidad de registros; 150,000, lo cual no es muy grande para nuestros fines, pero si representativo, pues tambien la página nos presenta un ejercicio con un data set de prueba, en la cual creemos que podemos ajustarnos bien alguna técnioca para lidiar con el desvalanceo. Si bien pudimos escoger un data set con mejores condiciones, el hambito financiero es de nuestro interes y la naturaleza de este data set, mas mas común y la principal dificultad.

# Datos
Los datos fueron obtenidos en Kaggle, pero para comodidad nuestra y del lector fueron cargadas en gitgub para poder importarlas directamente por medio del url del repositorio.

Se presentan 3 links, los cuales hacen referencia a;
  - `df_raw_training`,
  - `df_raw_test`,
  - `df_prob`.

  Son los tres archivos con datos dados por la página. Necesitamos entender estos archivos. Ahora procedemos a hacer la carga. Tambien adjutamos el arhivo en la carpeta de entrega.

  

import pandas as pd
import numpy as np

url_training = 'https://raw.githubusercontent.com/dt-sc-Josue-Diaz/preentrega/refs/heads/main/cs-training.csv'
df_raw_training = pd.read_csv(url_training)

url_test = 'https://raw.githubusercontent.com/dt-sc-Josue-Diaz/preentrega/refs/heads/main/cs-test.csv'
df_raw_test = pd.read_csv(url_test)

url_prob = 'https://raw.githubusercontent.com/dt-sc-Josue-Diaz/preentrega/refs/heads/main/sampleEntry.csv'
df_prob = pd.read_csv(url_prob )

df_raw_training.head()

df_raw_training.info()

df_raw_test.head()

df_raw_test.describe()

df_prob.head()

df_prob.describe()

Observamos que existe una columna que no tiene nombre, procedemos a renombrarla y además a asignar el `id` como índice para evitar tener redundancia en los datos.



def nombres_indices(df):
  if 'Unnamed: 0' in df.columns:
    df = df.rename(columns={'Unnamed: 0':'id'})
  if 'Id' in df.columns:
    df = df.rename(columns={'Id':'id'})
  if 'id' in df.columns:
    df.set_index('id', inplace=True)

  return df

df_raw_training = nombres_indices(df_raw_training )
df_raw_test = nombres_indices(df_raw_test)
df_prob = nombres_indices(df_prob)


df_raw_training.columns

Notamos que el archivo, de test es el dado por la página para realizar los entrenamientos, mientra que los otros dos funcionan para las evaluaciones, el último arhivo contiene probabilidadees de incumplimiento. Procedemos a tener un solo archivo

df_merged = nombres_indices(pd.merge(df_prob, df_raw_training, 
                                     left_index=True, 
                                     right_index=True,
                                     how = 'inner'))
df_merged.columns

Ahora procedemos a guardar este procesos en las carpetas correspondientes y tambien en github. Una vez que hagamos el push en gitdesktop podemos cargar desde internet para el segundo paso. Esto para fines prácticos, además que garantizamos que ambos datos son los mismos.


df_merged.to_csv('02_data_discovery/data_credit_test.csv', 
                 index = False)
df_raw_training.to_csv('02_data_discovery/data_para_discovery.csv', 
                       index = False)

# Bibliografia
https://carlosmarichal.colmex.mx/banca/Historia%20de%20la%20Banca%20y%20el%20credito%20en%20Mexico.pdf

https://www.asobancaria.com/wp-content/uploads/2018/01/13b-alberto-elices_2016-ajustes-de-valoracion-xva.pdf

https://www.burodecredito.com.mx/
