==> 02_data_discovery.ipynb <==
### Carga de datos
Una vez que tenemos los datos en formato csv, para poder manejarlos los cargamos anteriormene en github. Estos los importamos ahora mediante.

import pandas as pd
import numpy as np

url_data = 'https://raw.githubusercontent.com/dt-sc-Josue-Diaz/Proyecto-DS-II/refs/heads/main/02_data_discovery/data_para_discovery.csv'
df = pd.read_csv(url_data)


Para darnos la idea de que seguimos como en el primer paso, tenemos una muestra para observar que las configuraciones previas fueron las correctas.


df.sample(5)

# Valores nulos
Vamos a proceder con los valores nulos, necesitamos conocer el timpo de variables con datos nulos, a manera de sospecha, la falta de informacion siempre es causa de riesgo a la institución y a los procedimientos de concesion de créditos.



df.info()

De manera superficial notamos que si existen nulos, vemos ahora que la cantidad de datos es la que tenemos en la documentacion de la pagina de kaggle.

df.shape

Vamos a testear la cantidad porncentual de datos nulos.

(df.isnull().sum()/df.shape[0])*100

Notemos que tenemos una cantidad de nulos cercanas al 20%, pero como pista, la falta de información es causal de riesgo a la contraperte, por tal buscamos una manera de evitar eliminar tantos registros.

df[df['MonthlyIncome'].isna()].describe()

En contexto con los metadatos, observamos que DebRatio, es la proporción;

$$DebRatio=\frac{Gastos}{Ingresos}$$

se mide qué proporción del ingreso mensual de una persona se destina a cubrir deudas y gastos esenciales es decir, la capacidad que tiene una persona para solverntarse además, se puede entender como un indicador de la capacidad de endeudamiento de un individuo. Notemos que DebRatio mayor que uno nos está implicando que la persona no gana lo suficiente respecto a sus gastos o no está ganando dinero. De esta manera, una persona que por mas de 3 meses tenga un debratio mayor a uno tendría sentido pensar que no está generando ingresos.

df.loc[df.MonthlyIncome.isna() == True,'MonthlyIncome']

Hicismos una investigacion no desiva para conseguir el promedio de ingresos mensuales en el año de la información, para conseguir los datos siguientes

| **Rango de edad** | **Ingreso mensual promedio (USD)** |
|-------------------|------------------------------------|
| 16-24 años        | $1,800 - $2,200                    |
| 25-34 años        | $3,200 - $4,000                    |
| 35-44 años        | $4,500 - $5,500                    |
| 45-54 años        | $4,800 - $6,000                    |
| 55-64 años        | $4,200 - $5,000                    |
| 65 años o más     | $2,000 - $3,000                    |

De esta manera tendremos una idea de cuanto posiblemente ganaba una persona por edad en el 2011. Con eso definimos la funcion para poder asignar a cada registro un aproximado de lo que podieran ganar en ese entonces por grupo de edad.

def ingresos_promedio(age):
    # Respecto a la edad de la persona se le asigna un ingreso aproximado de ese año.
    if age < 25:
        return 2000
    elif age < 35:
        return 4500
    elif age < 45:
        return 4800
    else:
        return 4200

Pues bien, ahora notemos que; si el deberatio es menor que uno y no tiene ingresos mensuales entonces en verdad es una perdida de información por lo que asignarle el ingreso promedio por edad no sesgaria el estudio a diferencia de usar la media del data set. Pues como puede verse en la tabla anterior la edad es un factor en la cantidad de ingresos.

Por otro lado si el DebRatio es mayor que uno tambien tiene sentido pensar que la persona no tiene ingresos para al menos cubrir sus gastos y por tanto sus deudas. De esta manera asignamos a cero este ingreso.


df.loc[(df['DebtRatio'] <1) & (df['MonthlyIncome'].isna()), 'MonthlyIncome'] = df.loc[(df['DebtRatio'] < 1) & (df['MonthlyIncome'].isna()), 'age'].apply(ingresos_promedio).astype(float)
df.loc[(df['DebtRatio'] >= 1) & (df['MonthlyIncome'].isna()), 'MonthlyIncome'] = 0

df[df.DebtRatio ==0].describe()

df[df.DebtRatio>1].describe()

35137/df.shape[0]

df[df.MonthlyIncome.isna()].sum()

Estos son los casos minimos en los cuales podemos hacer cambios sin ser drásticos en los datos. Además observemos que;

(df.isnull().sum()/df.shape[0])*100


Ahora tenemos una pérdida de información del 2.616% Si no eliminamos estos datos. Se tiene en la variable target.

df['SeriousDlqin2yrs'].unique()

Porcentualmente tenemos;

df['SeriousDlqin2yrs'].value_counts()/df.shape[0]

Si eliminamos los datos que son NA, esto cambia a;

df = df[~df.NumberOfDependents.isna()]
df.reset_index(drop=True, inplace=True)

Lo cual nos permite no disminuir mas el objetivo de los impagos.

(df.isnull().sum()/df.shape[0])*100

df.NumberOfDependents.describe()

En este contexto creemos que no podemos tratar de imputar estos nulos, ya que la cantidad de personas por familia impactan mucho en el debratio. Por lo que agreagar datos podria sesgar de manera intencional un registro a una clase. Preferimos eliminarlos además que ahora solo es el 	2.616 porciento del total de la muestra y el impacto de la variable objetivo consideramos que es tolerable, tendremos esto en cuenta al momento del entrenamiento.

# Notas importantes

df.NumberOfDependents.mean()

df['NumberOfDependents'].value_counts()/df.shape[0]

sum(df.loc[df.SeriousDlqin2yrs==1,'age'] <= 25)

sum(df.loc[df.SeriousDlqin2yrs==1,'age'] >= 75)

### Duplicados

# A nivel general
df.duplicated().sum()

# A nivel PK
df.index.nunique()==df.shape[0]

## Nulos
Esta parte evalua si existen datos nulos.

df.isnull().sum()

No se observan datos nulos. Con eso se finaliza la integridad de los datos para el eda.

# Metadatos
Tenemos valores unicos en nuestro data set. Las siguientes lineas nos permitirámn hacer bien la estadística en la siguiente sección.

df.columns

## Diccionario de variales.

| Variable Name                                  | Description                                                                                         | Type      |
|------------------------------------------------|-----------------------------------------------------------------------------------------------------|-----------|
| SeriousDlqin2yrs                               | Person experienced 90 days past due delinquency or worse                                             | Y/N       |
| RevolvingUtilizationOfUnsecuredLines           | Total balance on credit cards and personal lines of credit except real estate divided by credit limits| percentage|
| age                                            | Age of borrower in years                                                                             | integer   |
| NumberOfTime30-59DaysPastDueNotWorse           | Number of times borrower has been 30-59 days past due but no worse in the last 2 years               | integer   |
| DebtRatio                                      | Monthly debt payments, alimony, living costs divided by monthly gross income                         | percentage|
| MonthlyIncome                                  | Monthly income                                                                                       | real      |
| NumberOfOpenCreditLinesAndLoans                | Number of open loans (installment like car loan or mortgage) and lines of credit (e.g. credit cards) | integer   |
| NumberOfTimes90DaysLate                        | Number of times borrower has been 90 days or more past due                                           | integer   |
| NumberRealEstateLoansOrLines                   | Number of mortgage and real estate loans including home equity lines of credit                       | integer   |
| NumberOfTime60-89DaysPastDueNotWorse           | Number of times borrower has been 60-89 days past due but no worse in the last 2 years               | integer   |
| NumberOfDependents                             | Number of dependents in family excluding themselves (spouse, children, etc.)                         | integer   |

Vamos a contextualizar la tabla anterior.


| Nombre de la variable| Descripción | Tipo|
|------------------------------------------------|-----------------------------------------------------------------------------------------------------|-----------|
| moroso |  Persona que no ha pagado en mas de 90 dias| Moroso/ No moroso  |
| Porcentaje crediticio | Proporcion de la linea de credito usada| Porcentaje|
|edad| Edad del cliente en años | entero |
|Retrasos bimestrales | Cantidad de veces que no se ha pagado en dos meses en una plazo de dos años | entero   |
| razon_de_deuda| Deudas dividido por el ingreso bruto | porcerntaje|
| ingreso_mensual| Ingreso mensual | decimal |
| cantidad_de_prestamos | Número de préstamos abiertos (a plazos, como préstamos para automóviles o hipotecas) y líneas de crédito (por ejemplo, tarjetas de crédito)| entero   |
|cartera_vencida| Número de veces que el prestatario ha estado vencido por 90 días o más| entero   |
| deuda_inmobiliara |Número de préstamos hipotecarios y inmobiliarios, incluidas líneas de crédito sobre el valor líquido de la vivienda| entero   |
| retrasos_trimestrales| Número de veces que el prestatario ha estado vencido entre 60 y 89 días pero no ha empeorado en los últimos 2 años| entero   |
|numero_dependientes | Cantidad de personas en la familia | entero   |

Por tanto vamos a renombrar para nuestro estudio.


df.columns = ['moroso', 'porcentaje_crediticio', 'edad', 'retrasos_bimestrales', 'razon_de_deuda','ingresos', 'cantidad_de_prestamos', 'cartera_vencida','deuda_inmobiliara','retrasos_trimestrales','numero_dependientes']

def categoria_edades(edad):
    # Aquí debes retornar un solo valor basado en la edad, por ejemplo:
    if edad < 35:
        return 'Joven'
    elif edad < 45:
        return 'Adulto Joven'
    elif edad < 65:
        return 'Adulto'
    else:
        return 'Adulto mayor'

Q1 = df['ingresos'].quantile(0.25)
Q2 = df['ingresos'].quantile(0.75)

# Esta fución depende totalmente de los datos. No pudimos conseguir informacion sobre esto, por lo que pensamos que es la mejor manera de segregar la información.
def clases_social(ingreso):
  if ingreso < Q1:
    return 'Bajo'
  elif ingreso < Q2:
    return 'Medio'
  else:
    return 'Alto'

def categoria(variable):
    # Respecto a la edad de la persona se le asigna un ingreso aproximado de ese año.
    if variable == 1 :
        return "Moroso"
    else:
        return "No moroso"

df['ingresos'].quantile(0.25),  df['ingresos'].quantile(0.75)

df['generacion'] = df.edad.apply(categoria_edades)
df['clase_social'] = df.ingresos.apply(clases_social)
df['moroso'] = df['moroso'].apply(categoria)
df['numero_dependientes'] = df['numero_dependientes'].astype(int)

df

df.to_csv('03_eda/data_para_eda.csv',index = False)
