==> 04_codificacion.ipynb <==
import pandas as pd

df = pd.read_csv('https://raw.githubusercontent.com/dt-sc-Josue-Diaz/Proyecto-DS-II/refs/heads/main/04_encoding/data_para_codificacion.csv')
df_copia = df.copy()

df_copia.columns

df_copia.drop(['generacion', 'clase_social'],axis='columns',
               inplace=True)

df_copia.head()

df_copia.to_csv("05_models_evaluation/data_para_entrenamiento.csv")
